[#471][#472] allow save, load, and autoload of configuration

+++ irods/test/helpers.py
@@ -7,6 +7,7 @@ import contextlib
 import shutil
 import hashlib
 import base64
+import logging
 import math
 import socket
 import inspect
@@ -15,7 +16,7 @@ import random
 import datetime
 import json
 import sys
-import logging
+import irods.client_configuration as config
 from irods.session import iRODSSession
 from irods.message import (iRODSMessage, IRODS_VERSION)
 from irods.password_obfuscation import encode
@@ -56,6 +57,14 @@ class iRODSUserLogins(object):
     def __del__(self):
         for username in self._users_to_remove:
             self.admin.users.remove(username)
+
+
+def configuration_file_exists():
+    try:
+        config.load(failure_modes = (config.NoConfigError,), logging_level = logging.DEBUG)
+    except config.NoConfigError:
+        return False
+    return True
 
 
 class StopTestsException(Exception):
@@ -307,14 +316,34 @@ def remove_unused_metadata(session):
 
 
 @contextlib.contextmanager
-def file_backed_up(filename):
-    with tempfile.NamedTemporaryFile(prefix=os.path.basename(filename)) as f:
-        shutil.copyfile(filename, f.name)
+def file_backed_up(filename, require_that_file_exists = True):
+    _basename = os.path.basename(filename) if os.path.exists(filename) else None
+    if _basename is None and require_that_file_exists:
+        err = RuntimeError("Attempted to back up a file which doesn't exist: %r" % (filename,))
+        raise err
+    with tempfile.NamedTemporaryFile(prefix=('tmp' if not _basename else _basename)) as f:
         try:
+            if _basename is not None:
+                shutil.copyfile(filename, f.name)
             yield filename
         finally:
-            shutil.copyfile(f.name, filename)
+            if _basename is None:
+                # Restore the condition of the file being absent.
+                if os.path.exists(filename):
+                    os.unlink(filename)
+            else:
+                # Restore the file's contents as they were originally.
+                shutil.copyfile(f.name, filename)
 
+@contextlib.contextmanager
+def environment_variable_backed_up(var):
+    old_value = os.environ.get(var)
+    try:
+        yield var
+    finally:
+        os.environ.pop(var,None)
+        if old_value is not None:
+            os.environ[var] = old_value
 
 def irods_session_host_local (sess):
     return socket.gethostbyname(sess.host) == \
